requirements.txt
ipywidgets==8.1.3
pillow==10.4.0
